+++ pcs/src/traits.rs
@@ -41,3 +41,36 @@ pub trait PolynomialCommitmentScheme {
         transcript: &mut Self::FiatShamirTranscript,
     ) -> bool;
 }
+
+pub(crate) trait TensorCodeIOPPCS {
+    fn codeword_len(&self) -> usize;
+
+    fn hamming_weight(&self) -> f64;
+
+    fn row_col_from_variables<F: Field>(num_vars: usize) -> (usize, usize) {
+        let elems_for_smallest_tree = tree::leaf_adic::<F>() * 2;
+
+        let row_num: usize = elems_for_smallest_tree;
+        let msg_size: usize = (1 << num_vars) / row_num;
+
+        (row_num, msg_size)
+    }
+
+    fn query_complexity(&self, soundness_bits: usize) -> usize {
+        // NOTE: use Ligero (AHIV22) or Avg-case dist to a code (BKS18)
+        // version of avg case dist in unique decoding technique.
+        let avg_case_dist = self.hamming_weight() / 3f64;
+        let sec_bits = -(1f64 - avg_case_dist).log2();
+
+        (soundness_bits as f64 / sec_bits).ceil() as usize
+    }
+
+    fn proximity_repetitions<F: Field>(&self, soundness_bits: usize) -> usize {
+        // NOTE: use Ligero (AHIV22) or Avg-case dist to a code (BKS18)
+        // version of avg case dist in unique decoding technique.
+        // Here is the probability union bound
+        let single_run_soundness_bits = F::FIELD_SIZE - self.codeword_len().ilog2() as usize;
+
+        (soundness_bits as f64 / single_run_soundness_bits as f64).ceil() as usize
+    }
+}
